Move compact_legacy to compaction.rs
